Extracted FilterPopup
